requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
diffusers>=0.24.0
trl>=0.7.0
accelerate>=0.25.0
datasets>=2.14.0
pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
pyyaml>=6.0
mlflow>=2.9.0
tensorboard>=2.15.0
pytest>=7.4.0
pytest-cov>=4.1.0
tqdm>=4.66.0
nltk>=3.8.0
pycocotools>=2.0.7
clip-by-openai>=1.0
sentencepiece>=0.1.99
sacrebleu>=2.3.0
rouge-score>=0.1.2
evaluate>=0.4.0
